--- span-normalizer/span-normalizer/src/main/java/org/hypertrace/core/spannormalizer/redaction/PIIPCIFieldType.java
@@ -1,16 +1,0 @@
-package org.hypertrace.core.spannormalizer.redaction;
-
-public enum PIIPCIFieldType {
-  PII("PII"),
-  PCI("PCI");
-
-  private final String value;
-
-  PIIPCIFieldType(String value) {
-    this.value = value;
-  }
-
-  public String getValue() {
-    return value;
-  }
-}
